validation ok

+++ src/routes/ConsentRoutes.ts
@@ -47,6 +47,46 @@ consentRoutes.get("/:id", async (req: Request, res: Response) => {
   });
 });
 
+consentRoutes.get(
+  "/document/:document",
+  async (req: Request, res: Response) => {
+    const { document } = req.params;
+    const consent = await prisma.consent.findFirst({
+      where: { userCPF: String(document) },
+      include: {
+        permissions: {
+          select: {
+            productName: true,
+          },
+        },
+      },
+    });
+
+    if (!consent) {
+      return res
+        .status(404)
+        .send({ message: "Consent not found with this document!" });
+    }
+
+    return res.status(200).send({
+      data: consent,
+      links: {
+        self: "https://65pwcy7ng5.execute-api.us-east-1.amazonaws.com//open-insurance/consents/v1",
+        first:
+          "https://65pwcy7ng5.execute-api.us-east-1.amazonaws.com//open-insurance/consents/v1",
+        prev: "https://65pwcy7ng5.execute-api.us-east-1.amazonaws.com//open-insurance/consents/v1",
+        next: "https://65pwcy7ng5.execute-api.us-east-1.amazonaws.com//open-insurance/consents/v1",
+        last: "https://65pwcy7ng5.execute-api.us-east-1.amazonaws.com//open-insurance/consents/v1",
+      },
+      meta: {
+        totalPages: 1,
+        totalRecords: 1,
+        requestDateTime: "2022-11-08T05:21:04.926Z",
+      },
+    });
+  }
+);
+
 consentRoutes.put("/:id", async (req: Request, res: Response) => {
   const { id } = req.params;
   const { status, permissions } = req.body;
@@ -91,14 +131,18 @@ consentRoutes.put("/:id", async (req: Request, res: Response) => {
 consentRoutes.delete("/:id", async (req: Request, res: Response) => {
   const { id } = req.params;
 
+  const test = await prisma.consent.findFirst({
+    where: { consentId: String(id) },
+  });
+
+  if (!test) {
+    return res.status(404).send({ message: "Consent not found!" });
+  }
   const consent = await prisma.consent.delete({
     where: { consentId: String(id) },
   });
 
-  if (!consent) {
-    return res.status(404).send({ message: "Consent not found!" });
-  }
-
+  console.log(consent);
   return res.status(200).send({
     message: "Consent with the ID = " + id + " deleted with success.",
   });
